requirements.txt
efficientunet-pytorch
matplotlib
nibabel
monai
tensorboardX
